Create Separate Certificate List Header component and code cleanup
